Activation/requirements.txt
torch>=2.0.0
transformers>=4.50.0,<4.60.0
datasets>=2.14.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0

